Add csrf_token

+++ src/Controller/PinsController.php
@@ -73,10 +73,14 @@ class PinsController extends AbstractController
     }
     #[Route('/pins/{id<[0-9]+>}/delete', name: 'app_pins_delete',methods:'GET|POST')]
 
-    public function delete ( Pin $pin,EntityManagerInterface $em): Response
+    public function delete (Request $request ,Pin $pin,EntityManagerInterface $em): Response
     {
-      $em ->remove($pin);
-      $em->flush();
+
+      if ($this->isCsrfTokenValid('pin_deletion_' .$pin->getId(),$request->request->get('Csrf_token'))) {
+         $em ->remove($pin);
+         $em->flush();
+      }
+
       return $this->redirectToRoute('app_home');
     }
 }
